% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{check_takuzu}
\alias{check_takuzu}
\title{Vérifie qu'une grille Takuzu est valide}
\usage{
check_takuzu(board)
}
\arguments{
\item{board}{Une matrice n x n de 0 et 1.}
}
\value{
TRUE si la grille est valide, FALSE sinon.
}
\description{
Cette fonction s'assure qu'une grille complète (sans NA) respecte les règles du Takuzu
}
