% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{generate_takuzu_solution}
\alias{generate_takuzu_solution}
\title{Génère une solution Takuzu complète}
\usage{
generate_takuzu_solution(n)
}
\arguments{
\item{n}{taille de la grille (doit être pair (ex. 6, 8).)}
}
\value{
Une matrice \code{n x n} avec de 0 et 1.
}
\description{
Tout est dans le titre
}
