% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{generateTakuzu}
\alias{generateTakuzu}
\title{Cette fonction génère une grille de Takuzu de taille \code{size} avec un pourcentage de remplissage spécifié par \code{fill_rate}.
Les cases de la grille sont initialisées à -1, et des valeurs 0 ou 1 sont placées aléatoirement en respectant les règles du jeu.}
\usage{
generateTakuzu(size, fill_rate)
}
\arguments{
\item{size}{Taille de la grille (un entier).}

\item{fill_rate}{Pourcentage de cases remplies (entre 0 et 1).}
}
\value{
Une matrice entière représentant la grille générée.
}
\description{
Cette fonction génère une grille de Takuzu de taille \code{size} avec un pourcentage de remplissage spécifié par \code{fill_rate}.
Les cases de la grille sont initialisées à -1, et des valeurs 0 ou 1 sont placées aléatoirement en respectant les règles du jeu.
}
