% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_grille_valide.R
\name{generer_grille_valide}
\alias{generer_grille_valide}
\title{Génère une grille de Takuzu valide}
\usage{
generer_grille_valide(nRows, nCols)
}
\arguments{
\item{nRows}{Nombre de lignes de la grille.}

\item{nCols}{Nombre de colonnes de la grille.}
}
\value{
Une matrice \code{nRows} x \code{nCols} remplie de "0" et "1" (caractères).
}
\description{
Cette fonction permet de générer une grille valide de Takuzu déjà remplie,
en respectant les règles du jeu (pas plus de deux chiffres identiques consécutifs,
répartition équilibrée des 0 et 1).
}
