% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generer_puzzle_valide}
\alias{generer_puzzle_valide}
\title{Génère une grille de jeu partiellement remplie}
\usage{
generer_puzzle_valide(taille, difficulte = "moyen")
}
\arguments{
\item{taille}{Taille de la grille}

\item{difficulte}{Niveau de difficulté : facile, moyen ou difficile}
}
\value{
Une liste avec la grille partielle et la solution
}
\description{
Génère une grille de jeu partiellement remplie
}
