% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generer_solution_valide}
\alias{generer_solution_valide}
\title{Génère une grille Takuzu complète et valide}
\usage{
generer_solution_valide(taille)
}
\arguments{
\item{taille}{Taille de la grille}
}
\value{
Une matrice complète
}
\description{
Génère une grille Takuzu complète et valide
}
