% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFillRate}
\alias{getFillRate}
\title{Get fill rate based on difficulty}
\usage{
getFillRate(difficulty)
}
\arguments{
\item{difficulty}{Game difficulty}
}
\value{
Fill rate
}
\description{
Get fill rate based on difficulty
}
