% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{grille}
\alias{grille}
\title{Génère une grille Takuzu 8x8 qui respecte les règles du Takuzu}
\usage{
grille(n)
}
\arguments{
\item{n}{Niveau de difficulté entre 'facile', 'moyen', 'difficile', 'jour'.
Le nombre de cases à supprimer dépend du niveau choisi.}
}
\value{
Une matrice 8x8 contenant la grille Takuzu avec certaines cases vides.
}
\description{
Génère une grille Takuzu 8x8 qui respecte les règles du Takuzu
}
