% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{has_three_consecutive}
\alias{has_three_consecutive}
\title{Détecte la présence de 3 valeurs identiques consécutives dans un vecteur}
\usage{
has_three_consecutive(x)
}
\arguments{
\item{x}{Un vecteur (ligne ou colonne).}
}
\value{
\code{TRUE} s'il y a 3 valeurs identiques consécutives, \code{FALSE} sinon.
}
\description{
Détecte la présence de 3 valeurs identiques consécutives dans un vecteur
}
\keyword{internal}
