% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{initializeGame}
\alias{initializeGame}
\title{Initialize new game}
\usage{
initializeGame(size, difficulty)
}
\arguments{
\item{size}{Grid size}

\item{difficulty}{Game difficulty}
}
\description{
Initialize new game
}
