% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{is_solved_takuzu}
\alias{is_solved_takuzu}
\title{Vérifie si une grille Takuzu est entièrement résolue}
\usage{
is_solved_takuzu(board)
}
\arguments{
\item{board}{Une matrice \code{n x n} de 0, 1 et éventuellement \code{NA}.}
}
\value{
\code{TRUE} si la grille est entièrement résolue, \code{FALSE} sinon.
}
\description{
Une grille est dite "résolue" si elle ne contient aucun \code{NA} et qu'elle
satisfait toutes les règles du Takuzu (voir \code{\link{check_takuzu}}).
}
