% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{is_valid_so_far}
\alias{is_valid_so_far}
\title{Vérifie si la case (row, col) est valide jusqu'à présent}
\usage{
is_valid_so_far(board, row, col)
}
\arguments{
\item{board}{Matrice partiellement remplie.}

\item{row}{Ligne de la case à vérifier.}

\item{col}{Colonne de la case à vérifier.}
}
\value{
\code{TRUE} si c'est valide, \code{FALSE} sinon.
}
\description{
Vérifie si la case (row, col) est valide jusqu'à présent
}
\keyword{internal}
