% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{is_valid_takuzu}
\alias{is_valid_takuzu}
\title{Is_valid_oui_ou_non}
\usage{
is_valid_takuzu(grid)
}
\value{
oui ou non
}
\description{
Est ce que la valeur rentré est juste ? cette fonction verifie ligne par ligne et colonne par colonne
}
