% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{isValidPlacement}
\alias{isValidPlacement}
\title{Check if placement is valid}
\usage{
isValidPlacement(grid, row, col, val)
}
\arguments{
\item{grid}{Current grid}

\item{row}{Row index}

\item{col}{Column index}

\item{val}{Value to check}
}
\value{
Logical
}
\description{
Check if placement is valid
}
