% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masquer_cases.R
\name{masquer_cases}
\alias{masquer_cases}
\title{Fonction qui permet de cacher des cases de la grille issue de la fonction \code{generer_grille_valide}}
\usage{
masquer_cases(grille, nb_vide)
}
\arguments{
\item{grille}{Grille valide de Takuzu (matrice de "0"/"1").}

\item{nb_vide}{Nombre de cases à cacher (entier).}
}
\value{
Une matrice avec des "0", "1" et des cases vides (chaînes vides).
}
\description{
Cette fonction prend une grille valide de Takuzu et remplace un certain
nombre de cases par des cases vides, afin de créer une grille jouable.
}
