% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Lancer l'application Takuzu}
\usage{
run_app()
}
\value{
L'application shiny composée du server et de l'ui définis au préalable.
}
\description{
Lancer l'application Takuzu
}
