% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{solveTakuzu}
\alias{solveTakuzu}
\title{Solve Takuzu grid}
\usage{
solveTakuzu(grid)
}
\arguments{
\item{grid}{Grille à résoudre (matrice d'entiers).}
}
\value{
La grille résolue ou \code{NULL} si aucune solution n'est trouvée.
}
\description{
Résout une grille de Takuzu en utilisant une approche récursive de recherche de solutions.
Retourne la grille résolue ou \code{NULL} si aucune solution n'est trouvée.
}
