% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{takuzu_feedback}
\alias{takuzu_feedback}
\title{Titre}
\usage{
takuzu_feedback(board)
}
\arguments{
\item{board}{Une matrice \code{n x n} de 0, 1 et éventuellement \code{NA}.}
}
\value{
Une matrice logique de dimension \code{n x n}.
}
\description{
Pour une grille partielle cette fonction renvoie une matrice de la même taille contenant \code{TRUE}
pour les cellules dont la valeur ne viole pas immédiatement les règles (dans
leur contexte de ligne et colonne) et \code{FALSE} pour les cellules qui posent problème.
}
