% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{takuzu_server}
\alias{takuzu_server}
\title{Takuzu Game Server Logic}
\usage{
takuzu_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Takuzu Game Server Logic
}
