% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verifier}
\alias{verifier}
\title{Vérifie que la grille générée respecte les règles du Takuzu.}
\usage{
verifier(g)
}
\arguments{
\item{g}{Une matrice 8x8 représentant la grille Takuzu.}
}
\value{
TRUE si la grille est valide, FALSE sinon.
}
\description{
Vérifie que la grille générée respecte les règles du Takuzu.
}
