% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglejeu.R
\name{verifier_grille}
\alias{verifier_grille}
\title{Vérifie si une grille de Takuzu est valide}
\usage{
verifier_grille(grille)
}
\arguments{
\item{grille}{Une matrice représentant la grille de jeu (0 et 1).}
}
\value{
TRUE si la grille est valide, FALSE sinon.
}
\description{
Vérifie si une grille de Takuzu est valide
}
