% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verifier_grille_partielle}
\alias{verifier_grille_partielle}
\title{Vérifie qu'une grille partiellement remplie est correcte}
\usage{
verifier_grille_partielle(grille, solution)
}
\arguments{
\item{grille}{La grille partielle}

\item{solution}{La solution}
}
\value{
TRUE/FALSE
}
\description{
Vérifie qu'une grille partiellement remplie est correcte
}
