% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verifier_solution_complete}
\alias{verifier_solution_complete}
\title{Vérifie si une grille est valide selon les règles du jeu}
\usage{
verifier_solution_complete(grille)
}
\arguments{
\item{grille}{Une matrice}
}
\value{
TRUE/FALSE
}
\description{
Vérifie si une grille est valide selon les règles du jeu
}
