% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{check_takuzu}
\alias{check_takuzu}
\title{Vérifie qu'une valeur dans le jeu de Takuzu est valide}
\usage{
check_takuzu(board)

check_takuzu(board)
}
\arguments{
\item{board}{Une matrice \code{n x n} de 0 et 1 (pas de \code{NA}).}
}
\value{
\code{TRUE} si la grille est valide, \code{FALSE} sinon.

oui ou non
}
\description{
Vérifie si la valeur placée dans la cellule (row, col) ne viole pas les règles du Takuzu.

pour le bouton verification
}
