% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{check_takuzu}
\alias{check_takuzu}
\title{Vérifie qu'une grille Takuzu est valide}
\usage{
check_takuzu(board)
}
\arguments{
\item{board}{Une matrice \code{n x n} de 0 et 1 (pas de \code{NA}).}
}
\value{
\code{TRUE} si la grille est valide, \code{FALSE} sinon.
}
\description{
Cette fonction s'assure qu'une grille complétée (sans \code{NA}) respecte
les règles du Takuzu : pas plus de 2 zéros/uns consécutifs, et autant de 0 que de 1
dans chaque ligne et colonne.
}
