% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{fill_board}
\alias{fill_board}
\title{Backtracking pour avoir une grille complète juste}
\usage{
fill_board(X, ligne, col)
}
\arguments{
\item{col}{Numéro de la colonne en cours.}

\item{board}{Matrice en cours de remplissage.}

\item{row}{Numéro de la ligne en cours.}
}
\value{
La matrice complétée si une solution est trouvée, ou \code{NULL} sinon.
}
\description{
#' Remplir récursivement la grille avec Backtracking
}
\keyword{internal}
