% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{fill_board}
\alias{fill_board}
\title{Fonction récursive de backtracking}
\usage{
fill_board(board, row, col)
}
\arguments{
\item{board}{Matrice en cours de remplissage.}

\item{row}{Numéro de la ligne en cours.}

\item{col}{Numéro de la colonne en cours.}
}
\value{
La matrice complétée si une solution est trouvée, ou \code{NULL} sinon.
}
\description{
Fonction récursive de backtracking
}
\keyword{internal}
