% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{generate_takuzu}
\alias{generate_takuzu}
\title{Génère un puzzle Takuzu}
\usage{
generate_takuzu(n = 6, difficulty = 0.5)
}
\arguments{
\item{n}{taille}

\item{difficulty}{proportion enlevé}
}
\value{
le plateau de jeu
}
\description{
On génère une grille complète et on enleve des cases pour jouer
}
