% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{is_valid_so_far}
\alias{is_valid_so_far}
\title{Vérifie la validité partielle d'une case dans une grille Takuzu}
\usage{
is_valid_so_far(board, row, col)
}
\arguments{
\item{board}{Matrice partiellement remplie.}

\item{row}{Ligne de la case à vérifier.}

\item{col}{Colonne de la case à vérifier.}
}
\value{
\code{TRUE} si c'est valide, \code{FALSE} sinon.
}
\description{
Vérifie si la valeur placée dans la cellule (row, col) ne viole pas les règles du Takuzu jusqu'à présent.
}
\keyword{internal}
