% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{takuzu_feedback}
\alias{takuzu_feedback}
\title{Test}
\usage{
takuzu_feedback(board)
}
\arguments{
\item{board}{Une matrice \code{n x n} de 0, 1 et éventuellement \code{NA}.}
}
\value{
Une matrice de dimension \code{n x n}.
}
\description{
fournir un " feedback " cellule par cellule sur la validité de la grille.
}
