% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{verifier_grille}
\alias{verifier_grille}
\title{Vérification de la validité d'une grille}
\usage{
verifier_grille(grille)
}
\arguments{
\item{grille}{Une matrice représentant la grille du jeu.}
}
\value{
TRUE si la grille est valide, FALSE sinon.
}
\description{
Vérifie que les règles du Takuzu sont respectées :
- Règle 1 : Chaque case de la grille doit être remplie avec un 0 ou un 1.
- Règle 2 : Chaque ligne et chaque colonne doivent contenir autant de 0 que de 1.
- Règle 3 : Il est interdit d’avoir trois 0 ou trois 1 consécutifs dans une ligne ou une colonne.
- Règle 4 : Deux lignes ou deux colonnes identiques sont interdites dans la même grille.
}
